requirements.txt
tensorflow==2.10.0
keras==2.10.0
flask==2.2.3
Werkzeug==2.3.7
numpy==1.23.5
pillow==9.4.0
matplotlib==3.7.1
scikit-learn==1.2.2